refactor: #21 - Adapter(Like, Search) 패키지 및 네이밍 변경
